Get python3 working again..
There is a change in the behavior of ``shutil.copytree`` between python2 and
python3.  In python2 it will copy symlinks by default (even though the
docstring seems to claim it won't) but it is more strictly enforced in the
python3 stdlib implementation.  This patch now explicitly argues
``symlinks=True`` which gets things working again on both python2 and python3.