Create community.yml
This file aims to provide a new area of contribution for a specific type of community collaborator. Until recently, there was not an area where contributors of this type could easily help with the Project, but the CommOps subgroup aims to help target this kind of collaborator, and more help is needed with these efforts. This is a beginning list of areas that contributors could check out to become involved.

There's not any new badges for the CommOps team yet, but this is subject to change.