Remove Rolekit
The most recent information on Rolekit I can find was in reference to Fedora 21, and with Cockpit having the functionality in Fedora 23. Since I can no longer find a project page, I find it suitable to remove Rolekit from the project prompts. Additionally, since I was unsure if Cockpit was the proper replacement for Rolekit, I have filed an issue on the topic.